Move service manager at the root of services
As this relies heavily on the services, it’s more logical to have it in
the same directory, to coordinate any child services.
