+++ internal/services/withoutexamples.go
@@ -1,6 +1,6 @@
 //go:build !withexamplebroker
 
-package manager
+package services
 
 // Stop is a no-op in production code.
 func (m *Manager) Stop() {}
